# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Calseta",
    "description": "SOC data platform for security agent consumption",
    "version": "dev"
  },
  "paths": {
...
    "/v1/indicators/{indicator_uuid}": {
      "get": {
        "tags": [
          "indicators"
        ],
        "summary": "Get Indicator",
        "description": "Return a single indicator with full enrichment data including raw\nprovider responses. Used for the indicator detail sheet drill-down.",
        "operationId": "get_indicator_v1_indicators__indicator_uuid__get",
        "parameters": [
          {
            "name": "indicator_uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Indicator Uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_IndicatorDetailResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "indicators"
        ],
        "summary": "Patch Indicator",
        "description": "Update an indicator's malice verdict.\n\nSend `{\"malice\": \"Malicious\"}` to set an analyst override.\nSend `{\"malice\": null}` to reset to the enrichment-computed value.",
        "operationId": "patch_indicator_v1_indicators__indicator_uuid__patch",
        "parameters": [
          {
            "name": "indicator_uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Indicator Uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IndicatorPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_IndicatorDetailResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DataResponse_IndicatorDetailResponse_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/IndicatorDetailResponse"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "DataResponse[IndicatorDetailResponse]"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IndicatorDetailResponse": {
        "properties": {
          "uuid": {
            "type": "string",
            "title": "Uuid"
          },
          "type": {
            "$ref": "#/components/schemas/IndicatorType"
          },
          "value": {
            "type": "string",
            "title": "Value"
          },
          "malice": {
            "type": "string",
            "title": "Malice"
          },
          "malice_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Malice Source"
          },
          "malice_overridden_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Malice Overridden At"
          },
          "first_seen": {
            "type": "string",
            "format": "date-time",
            "title": "First Seen"
          },
          "last_seen": {
            "type": "string",
            "format": "date-time",
            "title": "Last Seen"
          },
          "is_enriched": {
            "type": "boolean",
            "title": "Is Enriched"
          },
          "enrichment_results": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enrichment Results"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "uuid",
          "type",
          "value",
          "malice",
          "first_seen",
          "last_seen",
          "is_enriched",
          "created_at",
          "updated_at"
        ],
        "title": "IndicatorDetailResponse",
        "description": "Indicator with full enrichment data including raw provider responses."
      },
...
      "IndicatorPatch": {
        "properties": {
          "malice": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MaliceLevel"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "IndicatorPatch",
        "description": "Patch request for updating an indicator \u2014 PATCH /v1/indicators/{uuid}."
      },
...
      "IndicatorType": {
        "type": "string",
        "enum": [
          "ip",
          "domain",
          "hash_md5",
          "hash_sha1",
          "hash_sha256",
          "url",
          "email",
          "account"
        ],
        "title": "IndicatorType",
        "description": "Supported indicator-of-compromise (IOC) types."
      },
...
      "MaliceLevel": {
        "type": "string",
        "enum": [
          "Pending",
          "Benign",
          "Suspicious",
          "Malicious"
        ],
        "title": "MaliceLevel",
        "description": "Indicator malice verdict \u2014 shared validation enum."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}